--- src/Router/components/index.ts
@@ -1,3 +1,0 @@
-export { Route } from "./Route"
-export { RouteLink } from "./RouteLink"
-export { RouteSwitch } from "./RouteSwitch"
