+++ src/Router/hooks/useRouter.ts
@@ -2,10 +2,8 @@ import { useContext } from "react"
 
 import { RouterContext } from "../context/RouterContext"
 
-const useRouter = () => {
+export const useRouter = () => {
   const { path } = useContext(RouterContext)
 
   return [path]
 }
-
-export default useRouter
