+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 
-import { RouterProvider } from './Router/context/RouterProvider'
+import { RouterProvider } from './Router'
 import { DateProvider } from './common/contexts/DateContext'
 import App from './App'
